Fixes after upgrade engine from Abrodot to Godot

+++ Scenes/World/Camera/Shifts/ManualShake.cs
@@ -8,7 +8,7 @@ public class ManualShake : IShiftProvider
     /// <inheritdoc />
     public Vector2 Shift => IsAlive ? Rand.InsideUnitCircle * Strength : Vec();
 
-    public double Strength { get; set; } = 0;
+    public float Strength { get; set; } = 0;
 
     /// <inheritdoc />
     public bool IsAlive { get; set; } = true;
